Fix repaint call responsibility problem + small
When click the other place where should not be changed, the system
recognize that action as cancle action. So  set All of circle color to
black.
Fix showmovable when occurr Back Do at the outside of box from value -1
to 0. Because it occurring bug whenever get back do at the first
time or the later time.