Fix wrong number on left pieces satatus
When the piece on the board, the number of left piece of owener's should
be 0. So I change the logic to check if the piece is out of board or
not.
Change YutNoRiSet to fire event when the flags are change( flag for
game process and flag for player turn).